feat(config): added InferenceConfig to align with Tools
Added inference provider config fields to Config to align
with what we have for Tools.

The policy intersection is still hard to resolve.
Policies deal with attributes using inversion of control (local/remote),
but is also used to set or enforce configuration values (read-only, etc.).

The Config.Enforce method should be the place to deal with this,
however it's not possible to do it because at that level it has no knowledge of the
attributes of the different features (and this is difficult or impossible to change).

The alternative is to deal with Enabled/Disabled only at policy level and don't allow
to enable or disable globally by config.
A first step towards resolving this issue would be to define what can or can't be configured.
By having a clear Config struct and behaviors we can then decide on the right approach.