fix(log): disable logging to file by default
Currently, a file `ai-cli.log` is created automatically
whenever the user starts a chat session.

The logging feature is opt-out since the default CLI options
are to enable the log and provide a default logging filename.

This PR reverses the behavior so that logging is opt-in,
the file will be created only if the `--log-enabled` flag
is passed.